YASOD notebook * Add outlier detection WIP * calculate std deviation * calculat EMA and SMA with varying window sizes
As of ambiguous assessment definition, there is no obvious definition of
a 'datapoint' within the Challenge. So I have to wait for the reply of
apsa giving me more information regarding what feature should be used
for calculating MAs and taking 'datapoints' from. Taking binary values
for click does not make sense.